refactor prerenderer and add support for userData cleanup to prevent disk from filling up
